backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
anthropic==0.39.0
httpx==0.27.2
python-dotenv==1.0.1
pandas==2.2.3
